requirements/dev.txt
-r ./base.txt

django-extensions==3.0.5
pycodestyle==2.6.0
pytest==6.0.1
ipython
